requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.4
dnspython==2.7.0
email_validator==2.2.0
flake8==7.1.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
gitdb==4.0.11
gitflow==0.5.1
GitPython==3.1.43
idna==3.8
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.3
pytest-cov==6.0.0
python-dotenv==1.0.1
requests==2.32.3
smmap==5.0.1
urllib3==2.2.2
watchdog==4.0.2
Werkzeug==3.0.3
WTForms==3.1.2
